api_function_app/requirements.txt
azure-functions
requests
openai
